refactor folder structure for llm. Isolated langgraph components to its own folder.
