+++ backend/db/tabular/insert_table.py
@@ -1,4 +1,4 @@
-from utils.os_re_tools import remove_file_extension, create_folder_by_location, extract_table_name, save_uploaded_file
+from utilities.os_re_tools import remove_file_extension, create_folder_by_location, extract_table_name, save_uploaded_file
 from db.tabular.postgres_utilities import insert_csv_into_table, add_fuzzystrmatch_extension, generate_column_definitions
 from db.tabular.table_embeddings import get_docs_from_rows, create_embeddings_of_table_rows
 from config import postgres_var
