+++ backend/db/tabular/table_embeddings.py
@@ -1,5 +1,5 @@
 from config import postgres_var
-from llm_core.src.llm_utils.embedding_utils import get_embedder
+from llm_core.langgraph.utilities.embedding_utils import get_embedder
 from db.tabular.postgres_utilities import fetch_all_rows_from_table, create_langchain_docs_from_rows, get_all_columns_and_types
 from langchain_core.documents import Document
 from langchain_postgres import PGVector
