+++ backend/llm_core/langgraph/components/chains/chains.py
@@ -10,8 +10,7 @@ from langchain_community.agent_toolkits.sql.toolkit import SQLDatabaseToolkit
 from langchain_community.agent_toolkits import create_sql_agent
 
 from models.models import MessageState
-from rich import print as rprint
-from llm_core.src.prompt_engineering.templates import *
+from llm_core.langgraph.components.prompts.templates import *
 
 
 async def call_sql_agent(prompt:str, state: MessageState) -> MessageState:
