+++ backend/llm_core/llm_config.py
@@ -1,7 +1,6 @@
 import os
 from dotenv import load_dotenv
 
-
 load_dotenv()
 
 class LoadOpenAIConfig:
