+++ backend/routers/routes.py
@@ -11,12 +11,12 @@ from db.tabular.insert_pdf_record import ingest_pdf_into_postgres
 from db.tabular.pdf_record_operations import get_pdf_names_from_db, get_pdf_data
 
 # import os and task related functions
-from utils.os_re_tools import remove_file_extension, set_abs_path, if_path_exists
+from utilities.os_re_tools import remove_file_extension, set_abs_path, if_path_exists
 from services.tasks import delete_task_table
 
 # import llm related functions
-from llm_core.langgraph_stream import run_chatbots, active_websockets, tasks, manager, message_queue
-from llm_core.src.llm_utils.chatbot_manager import start_chatbot, set_table, set_pdf
+from llm_core.langgraph.services.chatbot.langgraph_stream import run_chatbots, active_websockets, tasks, manager, message_queue
+from llm_core.langgraph.services.chatbot.chatbot_manager import start_chatbot, set_table, set_pdf
 from models.models import TableNameRequest, PdfNameRequest, MessageInstance
